+++ release.config.js
@@ -6,7 +6,7 @@ module.exports = {
     [
       'semantic-release-rubygem',
       {
-        updateGemfileLock: 'bundle exec appraisal install',
+        updateGemfileLock: 'bundle exec appraisal bundle install',
       },
     ],
     [
@@ -22,4 +22,5 @@ module.exports = {
     ],
     '@semantic-release/github',
   ],
+  branches: ['main'],
 };