+++ spec/apollo-federation/entities_field_spec.rb
@@ -47,7 +47,7 @@ RSpec.describe ApolloFederation::EntitiesField do
       let(:type_with_key) do
         Class.new(base_object) do
           graphql_name 'TypeWithKey'
-          key fields: 'id'
+          key fields: :id
           field :id, 'ID', null: false
           field :other_field, 'String', null: true
         end
@@ -71,10 +71,8 @@ RSpec.describe ApolloFederation::EntitiesField do
 
         it 'sets the Query as the owner to the _entities field' do
           expect(
-            schema.graphql_definition
-              .types['Query']
+            schema.query
               .fields['_entities']
-              .metadata[:type_class]
               .owner.graphql_name,
           ).to eq('Query')
         end
@@ -233,7 +231,7 @@ RSpec.describe ApolloFederation::EntitiesField do
                 let(:type_with_key) do
                   Class.new(base_object) do
                     graphql_name 'TypeWithKey'
-                    key fields: 'id'
+                    key fields: :id
                     field :id, 'ID', null: false
                     field :other_field, 'String', null: false
 
@@ -284,7 +282,7 @@ RSpec.describe ApolloFederation::EntitiesField do
 
                     Class.new(base_object) do
                       graphql_name 'TypeWithKey'
-                      key fields: 'id'
+                      key fields: :id
                       field :id, 'ID', null: false
                       field :other_field, 'String', null: false
 
@@ -351,23 +349,25 @@ RSpec.describe ApolloFederation::EntitiesField do
     end
   end
 
-  context 'with the original runtime' do
-    it_behaves_like 'entities field' do
-      let(:base_schema) do
-        Class.new(GraphQL::Schema) do
-          include ApolloFederation::Schema
+  if Gem::Version.new(GraphQL::VERSION) < Gem::Version.new('1.12.0')
+    context 'with older versions of GraphQL and the interpreter runtime' do
+      it_behaves_like 'entities field' do
+        let(:base_schema) do
+          Class.new(GraphQL::Schema) do
+            use GraphQL::Execution::Interpreter
+            use GraphQL::Analysis::AST
+
+            include ApolloFederation::Schema
+          end
         end
       end
     end
   end
 
-  context 'with the interpreter runtime' do
+  if Gem::Version.new(GraphQL::VERSION) > Gem::Version.new('1.12.0')
     it_behaves_like 'entities field' do
       let(:base_schema) do
         Class.new(GraphQL::Schema) do
-          use GraphQL::Execution::Interpreter
-          use GraphQL::Analysis::AST
-
           include ApolloFederation::Schema
         end
       end